+++ src/net_db.c
@@ -443,7 +443,7 @@ netdbReloadState(void)
 	return;
     }
     t = buf = xcalloc(1, (size_t) sb.st_size + 1);
-    l = FD_READ_METHOD(fd, buf, sb.st_size);
+    l = FD_READ_METHOD(fd, buf, (int) sb.st_size);
     file_close(fd);
     if (l <= 0)
 	return;
